refactor: 블라인드 북 도메인 import 변경 #131
