+++ src/main/java/com/bookripple/api/domain/blindsalepost/service/BlindSalePostServiceImpl.java
@@ -1,8 +1,6 @@
 package com.bookripple.api.domain.blindsalepost.service;
 
-import com.bookripple.api.common.code.BlindSalePostErrorCode;
-import com.bookripple.api.common.code.CommonErrorCode;
-import com.bookripple.api.common.error.ApiException;
+import com.bookripple.api.domain.blindsalepost.code.BlindSalePostErrorCode;
 import com.bookripple.api.domain.blindsalepost.converter.BlindSalePostConverter;
 import com.bookripple.api.domain.blindsalepost.dto.BlindSalePostReqDto;
 import com.bookripple.api.domain.blindsalepost.dto.BlindSalePostResDto;
@@ -17,213 +15,222 @@ import com.bookripple.api.domain.book.repository.BookRepository;
 import com.bookripple.api.domain.member.entity.Member;
 import com.bookripple.api.domain.member.repository.MemberRepository;
 import com.bookripple.api.domain.order.repository.TradeRepository;
+import com.bookripple.api.global.code.CommonErrorCode;
+import com.bookripple.api.global.error.ApiException;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class BlindSalePostServiceImpl implements BlindSalePostService {
 
-    private final BlindSalePostRepository blindSalePostRepository;
-    private final MemberRepository memberRepository;
-    private final BookRepository bookRepository;
-    private final PurchaseRequestRepository purchaseRequestRepository;
-    private final TradeRepository tradeRepository;
+  private final BlindSalePostRepository blindSalePostRepository;
+  private final MemberRepository memberRepository;
+  private final BookRepository bookRepository;
+  private final PurchaseRequestRepository purchaseRequestRepository;
+  private final TradeRepository tradeRepository;
 
-    @Override
-    @Transactional
-    public BlindSalePostResDto.Create createPost(Long memberId, BlindSalePostReqDto.Create request) {
-        // 1. 등록할 회원과 실제 도서 정보를 조회합니다.
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new ApiException(CommonErrorCode.NOT_FOUND, "해당 회원을 찾을 수 없습니다."));
+  @Override
+  @Transactional
+  public BlindSalePostResDto.Create createPost(Long memberId, BlindSalePostReqDto.Create request) {
+    // 1. 등록할 회원과 실제 도서 정보를 조회합니다.
+    Member member = memberRepository.findById(memberId)
+        .orElseThrow(() -> new ApiException(CommonErrorCode.NOT_FOUND, "해당 회원을 찾을 수 없습니다."));
 
-        Book book = bookRepository.findById(request.actualBookId())
-                .orElseThrow(() -> new ApiException(CommonErrorCode.NOT_FOUND, "도서 정보를 찾을 수 없습니다."));
+    Book book = bookRepository.findById(request.actualBookId())
+        .orElseThrow(() -> new ApiException(CommonErrorCode.NOT_FOUND, "도서 정보를 찾을 수 없습니다."));
 
-        // 2. [Converter]를 사용하여 DTO를 엔티티로 변환합니다.
-        BlindSalePost post = BlindSalePostConverter.toBlindSalePost(request, member, book);
+    // 2. [Converter]를 사용하여 DTO를 엔티티로 변환합니다.
+    BlindSalePost post = BlindSalePostConverter.toBlindSalePost(request, member, book);
 
-        // 3. 변환된 게시글 엔티티를 DB에 저장합니다.
-        BlindSalePost savedPost = blindSalePostRepository.save(post);
+    // 3. 변환된 게시글 엔티티를 DB에 저장합니다.
+    BlindSalePost savedPost = blindSalePostRepository.save(post);
 
-        // 4. [Converter]를 사용하여 저장된 엔티티를 응답 DTO로 변환하여 반환합니다.
-        return BlindSalePostConverter.toCreateResponse(savedPost);
+    // 4. [Converter]를 사용하여 저장된 엔티티를 응답 DTO로 변환하여 반환합니다.
+    return BlindSalePostConverter.toCreateResponse(savedPost);
+  }
+
+  @Override
+  public BlindSalePostResDto.Detail getPostDetail(Long blindPostId) {
+    // 1. 게시글 존재 여부 확인
+    BlindSalePost post = blindSalePostRepository.findById(blindPostId)
+        .orElseThrow(() -> new ApiException(BlindSalePostErrorCode.POST_NOT_FOUND));
+
+    // 2. 해당 게시글에 들어온 요청 갯수만 카운트
+    long requestCount = purchaseRequestRepository.countByBlindSalePostId(blindPostId);
+
+    // 3. 상세 정보와 카운트만 반환
+    return BlindSalePostConverter.toDetail(post, requestCount);
+  }
+
+  @Override
+  public BlindSalePostResDto.PurchaseRequestList getPurchaseRequests(Long blindPostId) {
+    // 1. 게시글 존재 확인
+    if (!blindSalePostRepository.existsById(blindPostId)) {
+      throw new ApiException(BlindSalePostErrorCode.POST_NOT_FOUND);
+    }
+    // 2. 구매 요청자 명단만 조회
+    List<PurchaseRequest> requests = purchaseRequestRepository.findAllByBlindSalePostId(
+        blindPostId);
+
+    // 3. 명단 리스트 반환
+    return BlindSalePostConverter.toPurchaseRequestList(blindPostId, requests);
+  }
+
+  @Override
+  public BlindSalePostResDto.SliceResponse getMyPostList(Long memberId, PostStatus status,
+      Long cursor, int size) {
+    // 1. 다음 페이지 확인을 위해 size + 1개를 요청합니다.
+    Pageable pageable = PageRequest.of(0, size + 1);
+
+    // 2. 수정된 Repository 메서드 호출: 내 ID와 상태값으로 필터링합니다.
+    List<BlindSalePost> posts = (cursor == null)
+        ? blindSalePostRepository.findAllByMemberIdAndPostStatusOrderByIdDesc(memberId, status,
+        pageable)
+        : blindSalePostRepository.findAllByMemberIdAndPostStatusAndIdLessThanOrderByIdDesc(memberId,
+            status, cursor, pageable);
+
+    // 3. 무한 스크롤을 위한 다음 페이지 여부 계산
+    boolean hasNext = posts.size() > size;
+    if (hasNext) {
+      posts.remove(size);
     }
 
-    @Override
-    public BlindSalePostResDto.Detail getPostDetail(Long blindPostId) {
-        // 1. 게시글 존재 여부 확인
-        BlindSalePost post = blindSalePostRepository.findById(blindPostId)
-                .orElseThrow(() -> new ApiException(BlindSalePostErrorCode.POST_NOT_FOUND));
+    // 4. 마지막 아이템의 ID를 다음 커서로 지정
+    Long nextCursor = posts.isEmpty() ? null : posts.get(posts.size() - 1).getId();
 
-        // 2. 해당 게시글에 들어온 요청 갯수만 카운트
-        long requestCount = purchaseRequestRepository.countByBlindSalePostId(blindPostId);
+    // 5. [Converter]를 사용하여 DTO 리스트로 변환
+    List<BlindSalePostResDto.ListElement> content = posts.stream()
+        .map(BlindSalePostConverter::toListElement)
+        .collect(Collectors.toList());
 
-        // 3. 상세 정보와 카운트만 반환
-        return BlindSalePostConverter.toDetail(post, requestCount);
+    return BlindSalePostConverter.toSliceResponse(content, nextCursor, hasNext);
+  }
+
+  @Override
+  @Transactional
+  public void updatePost(Long memberId, Long blindBookId, BlindSalePostReqDto.Update request) {
+    // 1. 게시글 존재 여부 확인
+    BlindSalePost post = blindSalePostRepository.findById(blindBookId)
+        .orElseThrow(() -> new ApiException(BlindSalePostErrorCode.POST_NOT_FOUND));
+
+    // 2. 권한 확인: 본인의 글만 수정 가능
+    if (!post.getMember().getId().equals(memberId)) {
+      throw new ApiException(BlindSalePostErrorCode.NOT_POST_OWNER);
+    }
+    // 3. 비즈니스 검증: 예약 중이거나 판매 완료된 글은 수정 불가
+    if (post.getPostStatus() == PostStatus.RESERVED) {
+      throw new ApiException(BlindSalePostErrorCode.ALREADY_RESERVED);
+    }
+    if (post.getPostStatus() == PostStatus.SOLD_OUT) {
+      throw new ApiException(BlindSalePostErrorCode.ALREADY_SOLD_OUT);
     }
 
-    @Override
-    public BlindSalePostResDto.PurchaseRequestList getPurchaseRequests(Long blindPostId) {
-        // 1. 게시글 존재 확인
-        if (!blindSalePostRepository.existsById(blindPostId)) {
-            throw new ApiException(BlindSalePostErrorCode.POST_NOT_FOUND);
-        }
-        // 2. 구매 요청자 명단만 조회
-        List<PurchaseRequest> requests = purchaseRequestRepository.findAllByBlindSalePostId(blindPostId);
+    // 4. 엔티티의 update 메서드 호출 (Dirty Checking으로 자동 DB 반영)
+    post.update(
+        request.title(),
+        request.subtitle(),
+        request.description(),
+        request.price(),
+        BookCondition.valueOf(request.bookCondition())
+    );
+  }
 
-        // 3. 명단 리스트 반환
-        return BlindSalePostConverter.toPurchaseRequestList(blindPostId, requests);
+  @Override
+  @Transactional
+  public void deletePost(Long memberId, Long blindBookId) {
+    // 1. 삭제할 게시글 존재 확인
+    BlindSalePost post = blindSalePostRepository.findById(blindBookId)
+        .orElseThrow(() -> new ApiException(BlindSalePostErrorCode.POST_NOT_FOUND));
+
+    // 2. 권한 확인
+    if (!post.getMember().getId().equals(memberId)) {
+      throw new ApiException(BlindSalePostErrorCode.NOT_POST_OWNER);
     }
 
-    @Override
-    public BlindSalePostResDto.SliceResponse getMyPostList(Long memberId, PostStatus status, Long cursor, int size) {
-        // 1. 다음 페이지 확인을 위해 size + 1개를 요청합니다.
-        Pageable pageable = PageRequest.of(0, size + 1);
-
-        // 2. 수정된 Repository 메서드 호출: 내 ID와 상태값으로 필터링합니다.
-        List<BlindSalePost> posts = (cursor == null)
-                ? blindSalePostRepository.findAllByMemberIdAndPostStatusOrderByIdDesc(memberId, status, pageable)
-                : blindSalePostRepository.findAllByMemberIdAndPostStatusAndIdLessThanOrderByIdDesc(memberId, status, cursor, pageable);
-
-        // 3. 무한 스크롤을 위한 다음 페이지 여부 계산
-        boolean hasNext = posts.size() > size;
-        if (hasNext) {
-            posts.remove(size);
-        }
-
-        // 4. 마지막 아이템의 ID를 다음 커서로 지정
-        Long nextCursor = posts.isEmpty() ? null : posts.get(posts.size() - 1).getId();
-
-        // 5. [Converter]를 사용하여 DTO 리스트로 변환
-        List<BlindSalePostResDto.ListElement> content = posts.stream()
-                .map(BlindSalePostConverter::toListElement)
-                .collect(Collectors.toList());
-
-        return BlindSalePostConverter.toSliceResponse(content, nextCursor, hasNext);
+    // 3. 비즈니스 검증: 판매 완료된 글은 삭제 불가
+    if (post.getPostStatus() == PostStatus.SOLD_OUT) {
+      throw new ApiException(BlindSalePostErrorCode.ALREADY_SOLD_OUT);
     }
 
-    @Override
-    @Transactional
-    public void updatePost(Long memberId, Long blindBookId, BlindSalePostReqDto.Update request) {
-        // 1. 게시글 존재 여부 확인
-        BlindSalePost post = blindSalePostRepository.findById(blindBookId)
-                .orElseThrow(() -> new ApiException(BlindSalePostErrorCode.POST_NOT_FOUND));
+    // 3. 게시글 삭제 실행
+    blindSalePostRepository.delete(post);
+  }
 
-        // 2. 권한 확인: 본인의 글만 수정 가능
-        if (!post.getMember().getId().equals(memberId)) {
-            throw new ApiException(BlindSalePostErrorCode.NOT_POST_OWNER);
-        }
-        // 3. 비즈니스 검증: 예약 중이거나 판매 완료된 글은 수정 불가
-        if (post.getPostStatus() == PostStatus.RESERVED) {
-            throw new ApiException(BlindSalePostErrorCode.ALREADY_RESERVED);
-        }
-        if (post.getPostStatus() == PostStatus.SOLD_OUT) {
-            throw new ApiException(BlindSalePostErrorCode.ALREADY_SOLD_OUT);
-        }
+  // 1. 블라인드 북 전체 목록 (SALE 상태만)
+  @Override
+  public BlindSalePostResDto.BuyerSliceResponse<BlindSalePostResDto.BuyerListElement> getAllPosts(
+      Long cursor, int size) {
+    Pageable pageable = PageRequest.of(0, size + 1); // 다음 페이지 확인을 위해 하나 더 조회
 
-        // 4. 엔티티의 update 메서드 호출 (Dirty Checking으로 자동 DB 반영)
-        post.update(
-                request.title(),
-                request.subtitle(),
-                request.description(),
-                request.price(),
-                BookCondition.valueOf(request.bookCondition())
-        );
-    }
+    List<BlindSalePost> posts = (cursor == null)
+        ? blindSalePostRepository.findAllByPostStatusOrderByIdDesc(PostStatus.SALE, pageable)
+        : blindSalePostRepository.findAllByPostStatusAndIdLessThanOrderByIdDesc(PostStatus.SALE,
+            cursor, pageable);
 
-    @Override
-    @Transactional
-    public void deletePost(Long memberId, Long blindBookId) {
-        // 1. 삭제할 게시글 존재 확인
-        BlindSalePost post = blindSalePostRepository.findById(blindBookId)
-                .orElseThrow(() -> new ApiException(BlindSalePostErrorCode.POST_NOT_FOUND));
+    // 다음 페이지 존재 여부 확인
+    boolean hasNext = posts.size() > size;
+    List<BlindSalePost> contentPosts = hasNext ? posts.subList(0, size) : posts;
 
-        // 2. 권한 확인
-        if (!post.getMember().getId().equals(memberId)) {
-            throw new ApiException(BlindSalePostErrorCode.NOT_POST_OWNER);
-        }
+    // DTO 변환
+    List<BlindSalePostResDto.BuyerListElement> content = contentPosts.stream()
+        .map(BlindSalePostConverter::toBuyerListElement)
+        .collect(Collectors.toList());
 
-        // 3. 비즈니스 검증: 판매 완료된 글은 삭제 불가
-        if (post.getPostStatus() == PostStatus.SOLD_OUT) {
-            throw new ApiException(BlindSalePostErrorCode.ALREADY_SOLD_OUT);
-        }
+    // 다음 커서 결정 (마지막 요소의 ID)
+    Long nextCursor = hasNext ? contentPosts.get(size - 1).getId() : null;
 
-        // 3. 게시글 삭제 실행
-        blindSalePostRepository.delete(post);
-    }
+    return BlindSalePostConverter.toBuyerSlice(content, nextCursor, hasNext);
+  }
 
-    // 1. 블라인드 북 전체 목록 (SALE 상태만)
-    @Override
-    public BlindSalePostResDto.BuyerSliceResponse<BlindSalePostResDto.BuyerListElement> getAllPosts(Long cursor, int size) {
-        Pageable pageable = PageRequest.of(0, size + 1); // 다음 페이지 확인을 위해 하나 더 조회
+  // 2. 내가 요청한 책 목록 (오른쪽 탭)
+  @Override
+  public BlindSalePostResDto.BuyerSliceResponse<BlindSalePostResDto.MyRequestListElement> getMyRequests(
+      Long memberId, Long cursor, int size) {
+    Pageable pageable = PageRequest.of(0, size + 1);
 
-        List<BlindSalePost> posts = (cursor == null)
-                ? blindSalePostRepository.findAllByPostStatusOrderByIdDesc(PostStatus.SALE, pageable)
-                : blindSalePostRepository.findAllByPostStatusAndIdLessThanOrderByIdDesc(PostStatus.SALE, cursor, pageable);
+    List<PurchaseRequest> requests = (cursor == null)
+        ? purchaseRequestRepository.findAllByMemberIdOrderByIdDesc(memberId, pageable)
+        : purchaseRequestRepository.findAllByMemberIdAndIdLessThanOrderByIdDesc(memberId, cursor,
+            pageable);
 
-        // 다음 페이지 존재 여부 확인
-        boolean hasNext = posts.size() > size;
-        List<BlindSalePost> contentPosts = hasNext ? posts.subList(0, size) : posts;
+    boolean hasNext = requests.size() > size;
+    List<PurchaseRequest> contentRequests = hasNext ? requests.subList(0, size) : requests;
 
-        // DTO 변환
-        List<BlindSalePostResDto.BuyerListElement> content = contentPosts.stream()
-                .map(BlindSalePostConverter::toBuyerListElement)
-                .collect(Collectors.toList());
+    // DTO 변환 (각 PurchaseRequest에 대한 Trade 조회)
+    List<BlindSalePostResDto.MyRequestListElement> content = contentRequests.stream()
+        .map(request -> {
+          Long tradeId = tradeRepository
+              .findByBuyerIdAndBlindSalePostId(memberId, request.getBlindSalePost().getId())
+              .map(trade -> trade.getId())
+              .orElse(null);
+          return BlindSalePostConverter.toMyRequestListElement(request, tradeId);
+        })
+        .collect(Collectors.toList());
 
-        // 다음 커서 결정 (마지막 요소의 ID)
-        Long nextCursor = hasNext ? contentPosts.get(size - 1).getId() : null;
+    // 다음 커서 결정 (마지막 요소의 ID)
+    Long nextCursor = hasNext ? contentRequests.get(contentRequests.size() - 1).getId() : null;
 
-        return BlindSalePostConverter.toBuyerSlice(content, nextCursor, hasNext);
-    }
+    return BlindSalePostConverter.toBuyerSlice(content, nextCursor, hasNext);
+  }
 
-    // 2. 내가 요청한 책 목록 (오른쪽 탭)
-    @Override
-    public BlindSalePostResDto.BuyerSliceResponse<BlindSalePostResDto.MyRequestListElement> getMyRequests(Long memberId, Long cursor, int size) {
-        Pageable pageable = PageRequest.of(0, size + 1);
+  @Override
+  @Transactional(readOnly = true)
+  public BlindSalePostResDto.BuyerDetail getPostDetailForBuyer(Long memberId, Long blindPostId) {
 
-        List<PurchaseRequest> requests = (cursor == null)
-                ? purchaseRequestRepository.findAllByMemberIdOrderByIdDesc(memberId, pageable)
-                : purchaseRequestRepository.findAllByMemberIdAndIdLessThanOrderByIdDesc(memberId, cursor, pageable);
+    BlindSalePost post = blindSalePostRepository.findById(blindPostId)
+        .orElseThrow(() -> new ApiException(BlindSalePostErrorCode.POST_NOT_FOUND));
 
-        boolean hasNext = requests.size() > size;
-        List<PurchaseRequest> contentRequests = hasNext ? requests.subList(0, size) : requests;
+    Optional<PurchaseRequest> myRequest = purchaseRequestRepository
+        .findByBlindSalePostIdAndMemberId(blindPostId, memberId);
 
-        // DTO 변환 (각 PurchaseRequest에 대한 Trade 조회)
-        List<BlindSalePostResDto.MyRequestListElement> content = contentRequests.stream()
-                .map(request -> {
-                    Long tradeId = tradeRepository
-                            .findByBuyerIdAndBlindSalePostId(memberId, request.getBlindSalePost().getId())
-                            .map(trade -> trade.getId())
-                            .orElse(null);
-                    return BlindSalePostConverter.toMyRequestListElement(request, tradeId);
-                })
-                .collect(Collectors.toList());
-
-        // 다음 커서 결정 (마지막 요소의 ID)
-        Long nextCursor = hasNext ? contentRequests.get(contentRequests.size() - 1).getId() : null;
-
-        return BlindSalePostConverter.toBuyerSlice(content, nextCursor, hasNext);
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public BlindSalePostResDto.BuyerDetail getPostDetailForBuyer(Long memberId, Long blindPostId) {
-
-        BlindSalePost post = blindSalePostRepository.findById(blindPostId)
-                .orElseThrow(() -> new ApiException(BlindSalePostErrorCode.POST_NOT_FOUND));
-
-        Optional<PurchaseRequest> myRequest = purchaseRequestRepository
-                .findByBlindSalePostIdAndMemberId(blindPostId, memberId);
-
-        return BlindSalePostConverter.toBuyerDetail(post, myRequest);
-    }
+    return BlindSalePostConverter.toBuyerDetail(post, myRequest);
+  }
 }
